+++ docs/CHANGE_LOG.md
@@ -1,6 +1,13 @@
 # Tree.js - Change Log:
 
 ## Version 0.2.0:
+
+#### **Binding Options:**
+- Added a new binding option called "showChildrenToggle", which states if the "Show Children" toggle check box should be shown (defaults to true).
+- Added a new binding option called "showDescriptionsToggle", which states if the "Show Descriptions" toggle check box should be shown (defaults to true).
+- Added a new binding option called "showContentsToggle", which states if the "Show Contents" toggle check box should be shown (defaults to true).
+
+#### **Documentation:**
 - Fixed a grammar mistake in the projects description.
 
 <br>